+++ testapps/ghissues/src/main/java/jpa_common/AppEntry.java
@@ -1,0 +1,11 @@
+package jpa_common;
+
+import act.Act;
+
+public class AppEntry {
+
+    public static void main(String[] args) throws Exception {
+        Act.start();
+    }
+
+}