+++ testapps/ghissues/src/main/java/jpa_common/gh31/Pageable.java
@@ -1,0 +1,7 @@
+package jpa_common.gh31;
+
+import act.util.SimpleBean;
+
+public class Pageable implements SimpleBean {
+    public String order;
+}